File structure reorganization.
