+++ ipymd/__init__.py
@@ -1,3 +1,3 @@
 from . import formats
-from .core import convert, format_manager
-from .contents_manager import *
+from .core.core import convert, format_manager
+from .core.contents_manager import *
