+++ ipymd/formats/markdown.py
@@ -13,8 +13,8 @@ Much of the code comes from the mistune library.
 import re
 from collections import OrderedDict
 
-from ..six import StringIO
-from ..utils import _ensure_string, _preprocess
+from ..ext.six import StringIO
+from ..utils.utils import _ensure_string, _preprocess
 
 
 #------------------------------------------------------------------------------
